+++ stdlib/LinearAlgebra/test/qr.jl
@@ -77,9 +77,7 @@ rectangularQ(Q::LinearAlgebra.AbstractQ) = convert(Array, Q)
                 @test Base.propertynames(qra)       == (:R, :Q)
             end
             @testset "Thin QR decomposition (without pivoting)" begin
-                # test inference within function calls, otherwise const prop won't work
-                qrnone(A) = qr(A, :none)
-                qra   = @inferred qrnone(a[:, 1:n1])
+                qra   = @inferred qr(a[:, 1:n1], NoPivot())
                 q,r   = qra.Q, qra.R
                 @test_throws ErrorException qra.Z
                 @test q'*squareQ(q) ≈ Matrix(I, a_1, a_1)
@@ -104,9 +102,7 @@ rectangularQ(Q::LinearAlgebra.AbstractQ) = convert(Array, Q)
                 @test Base.propertynames(qra)       == (:R, :Q)
             end
             @testset "(Automatic) Fat (pivoted) QR decomposition" begin
-                qrcolnorm(A) = qr(A, :colnorm)
-                # test inference within function calls
-                @inferred qrcolnorm(a)
+                @inferred qr(a, ColNorm())
 
                 qrpa  = factorize(a[1:n1,:])
                 q,r = qrpa.Q, qrpa.R
@@ -192,7 +188,7 @@ rectangularQ(Q::LinearAlgebra.AbstractQ) = convert(Array, Q)
                 @test mul!(c, b, q') ≈ b*q'
                 @test_throws DimensionMismatch mul!(Matrix{eltya}(I, n+1, n), q, b)
 
-                qra = qr(a[:,1:n1], :none)
+                qra = qr(a[:,1:n1], NoPivot())
                 q, r = qra.Q, qra.R
                 @test rmul!(copy(squareQ(q)'), q) ≈ Matrix(I, n, n)
                 @test_throws DimensionMismatch rmul!(Matrix{eltya}(I, n+1, n+1),q)
@@ -217,8 +213,8 @@ end
 @testset "transpose errors" begin
     @test_throws MethodError transpose(qr(randn(3,3)))
     @test_throws MethodError adjoint(qr(randn(3,3)))
-    @test_throws MethodError transpose(qr(randn(3,3), :none))
-    @test_throws MethodError adjoint(qr(randn(3,3), :none))
+    @test_throws MethodError transpose(qr(randn(3,3), NoPivot()))
+    @test_throws MethodError adjoint(qr(randn(3,3), NoPivot()))
     @test_throws MethodError transpose(qr(big.(randn(3,3))))
     @test_throws MethodError adjoint(qr(big.(randn(3,3))))
 end
@@ -258,7 +254,7 @@ end
     A = zeros(1, 2)
     B = zeros(1, 1)
     @test A \ B == zeros(2, 1)
-    @test qr(A, :colnorm) \ B == zeros(2, 1)
+    @test qr(A, ColNorm()) \ B == zeros(2, 1)
 end
 
 @testset "Issue 24107" begin
@@ -280,7 +276,7 @@ end
     @test A \b ≈ ldiv!(c, qr(A ), b)
     @test b == b0
     c0 = copy(c)
-    @test Ac\c ≈ ldiv!(b, qr(Ac, :colnorm), c)
+    @test Ac\c ≈ ldiv!(b, qr(Ac, ColNorm()), c)
     @test c0 == c
 end
 
@@ -297,7 +293,7 @@ end
 
 @testset "det(Q::Union{QRCompactWYQ, QRPackedQ})" begin
     # 40 is the number larger than the default block size 36 of QRCompactWY
-    @testset for n in [1:3; 40], m in [1:3; 40], pivot in (:none, :colnorm)
+    @testset for n in [1:3; 40], m in [1:3; 40], pivot in (NoPivot(), ColNorm())
         @testset "real" begin
             @testset for k in 0:min(n, m, 5)
                 A = cat(Array(I(k)), randn(n - k, m - k); dims=(1, 2))